LabService/backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.1
pyodbc==5.1.0
python-dotenv==1.0.1
pydantic==2.7.4
pydantic-settings==2.4.0
